(pure-move) migrate pad & unpad
